requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
aiosignal==1.2.0
aiosqlite==0.17.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.0
httptools==0.4.0
idna==3.3
iso8601==1.0.2
loguru==0.6.0
multidict==6.0.2
pydantic==1.9.1
pypika-tortoise==0.1.6
pyTelegramBotAPI==4.6.0
pytz==2022.1
PyYAML==6.0
requests==2.28.1
sanic==22.6.0
sanic-ext==22.6.2
sanic-routing==22.3.0
tortoise-orm==0.19.2
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.10
uvloop==0.16.0
websockets==10.3
yarl==1.7.2
